Implement RTCPeerConnection.restartIce() according to spec.
This is part of "Perfect Negotiation" (https://crbug.com/980872).
Spec: https://w3c.github.io/webrtc-pc/#dom-rtcpeerconnection-restartice

Intent to Ship:
https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/OqPfCpC5RYU

The restartIce() makes the next createOffer() generate new ICE
credentials, as if {iceRestart:true} was passed in as options. It also
causes negotiationneeded. This is better than manually restarting ICE
because it survives rollbacks (when that is implemented) and
restartIce() can be called regardless of current signalingState.
Problems with {restartIce:true} is described here:
https://docs.google.com/presentation/d/1xcvf0udNeSH7s1FOY7RRqr1dEFvokZjn-MZPjwy3iXQ/edit#slide=id.g5c2f3df65b_11_574

Bug: chromium:980881
Change-Id: Iaeb32d39d3ec514f19eb95150c3df257bcb5c754