Handle display:contents for viewport propagation.
When documentElement/body doe not generate a box, its background will
not propagate to the viewport (see [1]). This removes the need
for calling EnsureComputedStyle() on documentElement/body.

The CSSWG has also resolved that to be the case also for other
properties propagated to the viewport[2].

Some of the wpt tests for body propagation didn't actually have a body
element. Added.

[1] https://drafts.csswg.org/css-backgrounds/#special-backgrounds
[2] https://github.com/w3c/csswg-drafts/issues/3779#issuecomment-481762912

Bug: 987207

Change-Id: I06e618e2acd2926b5ae4831bf5825e13e970d035